--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "perks"]
-	path = perks
-	url = https://github.com/azure/perks
