+++ .scripts/npm-run.js
@@ -1,0 +1,2 @@
+// Runs the npm run command on each project that has it.
+require('./for-each').npm(process.argv[2]);
