+++ common/config/rush/command-line.json
@@ -5,44 +5,49 @@
       "commandKind": "bulk",
       "name": "set-version",
       "summary": "set the version in the package.json files",
-      "enableParallelism": true,
-      "ignoreMissingScript": true
+      "enableParallelism": true
+    },
+    {
+      "commandKind": "bulk",
+      "name": "reset-version",
+      "summary": "reset the version in the package.json files to .0",
+      "enableParallelism": true
     },
     {
       "commandKind": "global",
       "name": "sync-versions",
       "summary": "sync versions of sibling projects",
-      "shellCommand": "node common/scripts/sync-versions.js"
+      "shellCommand": "node .scripts/sync-versions.js"
     },
     {
       "commandKind": "global",
       "name": "watch",
       "summary": "run npm watch on all projects",
-      "shellCommand": "node common/scripts/watch.js"
+      "shellCommand": "node .scripts/watch.js"
     },
     {
       "commandKind": "global",
       "name": "clean",
       "summary": "run npm clean on all projects",
-      "shellCommand": "node common/scripts/clean.js"
-    },
-    {
-      "commandKind": "global",
-      "name": "fix",
-      "summary": "run all npm fix",
-      "shellCommand": "node common/scripts/fix.js"
-    },
-    {
-      "commandKind": "global",
-      "name": "lint",
-      "summary": "run all npm eslint",
-      "shellCommand": "node common/scripts/eslint.js"
+      "shellCommand": "node .scripts/npm-run.js clean"
     },
     {
       "commandKind": "global",
       "name": "test",
       "summary": "run all npm test",
-      "shellCommand": "node common/scripts/test.js"
+      "shellCommand": "node .scripts/npm-run.js test"
+    },
+    {
+      "commandKind": "global",
+      "name": "fix",
+      "summary": "run all npm fix",
+      "shellCommand": "node .scripts/npm-run.js eslint-fix"
+    },
+    {
+      "commandKind": "global",
+      "name": "lint",
+      "summary": "run all npm lint",
+      "shellCommand": "node .scripts/npm-run.js eslint"
     }
   ],
   "parameters": []
