--- common/scripts/clean.js
@@ -1,1 +1,0 @@
-require('./for-each').npm('clean');
