--- common/scripts/fix.js
@@ -1,1 +1,0 @@
-require('./for-each').npm('eslint-fix');
